Refactor pricing to use decimal for accuracy
Changed all price and total representations from double/int to decimal in Product, Checkout, ICheckout, and related tests to improve monetary precision and correctness.

+++ CheckoutLogic.Tests/Checkout_Scan_Tests.cs
@@ -14,7 +14,7 @@ public class Checkout_Scan_Tests
     [InlineData(new string[] { "A", "B" }, 80)]
     [InlineData(new string[] { "A", "B", "C" }, 100)]
     [InlineData(new string[] { "A", "B", "C", "D" }, 115)]
-    public void Scan_MultipleDifferentItems_ReturnsSumOfPrices(string[] items, double expectedTotal)
+    public void Scan_MultipleDifferentItems_ReturnsSumOfPrices(string[] items, decimal expectedTotal)
     {
         Checkout co = CreateCheckout();
