+++ CheckoutLogic/Models/Product.cs
@@ -1,8 +1,8 @@
 ﻿
 namespace CheckoutLogic.Models;
 
-public class Product(string SKU, double UnitPrice)
+public class Product(string SKU, decimal UnitPrice)
 {
     public string SKU { get; } = SKU;
-    public double UnitPrice { get; } = UnitPrice;
+    public decimal UnitPrice { get; } = UnitPrice;
 }
